Simplify pivot subsystem
After a discussion with Nolan, we both felt the Pivot subsystem was getting too
complicated to debug. This code simplifies the Pivot. It also changes the "rotation"
term to "angle" to be more specific, as it's actually an angle in degrees.

This also completes the PivotZero command which works correctly and is tested.